Change to better distinguish signature schemes
It is currently possible to call `APK::verify` or `PublicKey::verify` on
the same signature type, leading to situations where even if a user is
technically *allowed* to call the function, verification is in principle
impossible, since the signature schemes used to produce a signature are
different. Furthermore, the `sign_vulnerable` function is arguably
misnamed, since it is only actually vulnerable to a rogue key attack
*when used in a multi-signature context*.

In the interest of fixing this, this commit introduces a couple of changes
to the API of the crate, designed to better distinguish and clarify the
usage of the two different signature schemes it provides.

First and foremost, a new signature type - `MultisigSignature` is added,
representing a signature produced using the multi-signature scheme.
Conversely, the already existing `Signature` type is taken to mean the a
signature produced using the single-key scheme.
The functions under `SecretKey`, `sign_vulnerable` and `sign`, are
respectively renamed into `sign` and `sign_multisig`, and the latter is
changed to return the new `MultisigSignature` type.
The "aggregated public key" type is renamed from `APK` into
`MultisigPublicKey`, to better denote its use and to be in line with the
kind signature it can now verify using the `verify` function - which is
changed to take a `MultisigSignature`.

On a more subtle note, the `From<&PublicKey>` and `From<&SecretKey>`
implementations for `MultisigPublicKey`, formerly `APK` are removed and
the API of `MultisigPublicKey::aggregate` is changed to *not* take
`&self`, and only take a slice of `PublicKey`. This is designed to
promote the intended usage of the struct - aggregating a collection of
public keys, as opposed to just "aggregating" one.

Resolves: #18